// src/validator.d.ts
export declare class RegExpValidator {
    private readonly _options;
    private readonly _reader;
    private _uFlag;
    private _nFlag;
    private _lastIntValue;
    private _lastMinValue;
    private _lastMaxValue;
    private _lastStrValue;
    private _lastKeyValue;
    private _lastValValue;
    private _lastAssertionIsQuantifiable;
    private _numCapturingParens;
    private _groupNames;
    private _backreferenceNames;
    /**
     * Initialize this validator.
     * @param options The options of validator.
     */
    constructor(options?: RegExpValidator.Options);
    /**
     * Validate a regular expression literal. E.g. "/abc/g"
     * @param source The source code to validate.
     * @param start The start index in the source code.
     * @param end The end index in the source code.
     */
    validateLiteral(source: string, start?: number, end?: number): void;
    /**
     * Validate a regular expression flags. E.g. "gim"
     * @param source The source code to validate.
     * @param start The start index in the source code.
     * @param end The end index in the source code.
     */
    validateFlags(source: string, start?: number, end?: number): void;
    /**
     * Validate a regular expression pattern. E.g. "abc"
     * @param source The source code to validate.
     * @param start The start index in the source code.
     * @param end The end index in the source code.
     * @param uFlag The flag to set unicode mode.
     */
    validatePattern(source: string, start?: number, end?: number, uFlag?: boolean): void;
    private readonly strict;
    private readonly ecmaVersion;
    private onLiteralEnter;
    private onLiteralLeave;
    private onFlags;
    private onPatternEnter;
    private onPatternLeave;
    private onDisjunctionEnter;
    private onDisjunctionLeave;
    private onAlternativeEnter;
    private onAlternativeLeave;
    private onGroupEnter;
    private onGroupLeave;
    private onCapturingGroupEnter;
    private onCapturingGroupLeave;
    private onQuantifier;
    private onLookaroundAssertionEnter;
    private onLookaroundAssertionLeave;
    private onEdgeAssertion;
    private onWordBoundaryAssertion;
    private onAnyCharacterSet;
    private onEscapeCharacterSet;
    private onUnicodePropertyCharacterSet;
    private onCharacter;
    private onBackreference;
    private onCharacterClassEnter;
    private onCharacterClassLeave;
    private onCharacterClassRange;
    private readonly source;
    private readonly index;
    private readonly currentCodePoint;
    private readonly nextCodePoint;
    private readonly nextCodePoint2;
    private readonly nextCodePoint3;
    private reset;
    private rewind;
    private advance;
    private eat;
    private eat2;
    private eat3;
    private raise;
    private eatRegExpBody;
    private pattern;
    private countCapturingParens;
    private disjunction;
    private alternative;
    private eatTerm;
    private eatAssertion;
    private eatQuantifier;
    private eatBracedQuantifier;
    private eatAtom;
    private eatDot;
    private eatReverseSolidusAtomEscape;
    private eatUncapturingGroup;
    private eatCapturingGroup;
    private eatExtendedAtom;
    private eatReverseSolidusFollowedByC;
    private eatInvalidBracedQuantifier;
    private eatSyntaxCharacter;
    private eatPatternCharacter;
    private eatExtendedPatternCharacter;
    private groupSpecifier;
    private eatGroupName;
    private eatRegExpIdentifierName;
    private eatRegExpIdentifierStart;
    private eatRegExpIdentifierPart;
    private eatAtomEscape;
    private eatBackreference;
    private eatKGroupName;
    private eatCharacterEscape;
    private eatCControlLetter;
    private eatZero;
    private eatControlEscape;
    private eatControlLetter;
    private eatRegExpUnicodeEscapeSequence;
    private eatIdentityEscape;
    private isValidIdentityEscape;
    private eatDecimalEscape;
    private eatCharacterClassEscape;
    private eatUnicodePropertyValueExpression;
    private eatUnicodePropertyName;
    private eatUnicodePropertyValue;
    private eatLoneUnicodePropertyNameOrValue;
    private eatCharacterClass;
    private classRanges;
    private eatClassAtom;
    private eatClassEscape;
    private eatClassControlLetter;
    private eatHexEscapeSequence;
    private eatDecimalDigits;
    private eatHexDigits;
    private eatLegacyOctalEscapeSequence;
    private eatOctalDigit;
    private eatFixedHexDigits;
}